FRONTEND-1 feat: add filter menu
